feat: add getBookById and Id validation

+++ src/controllers/book.js
@@ -1,4 +1,5 @@
 const BookService = require('../services/Book');
+const mongoose = require('mongoose');
 
 module.exports.getAll = async(req, res) => {
   try {
@@ -9,6 +10,20 @@ module.exports.getAll = async(req, res) => {
   } catch (error) {
     res.status(500).json({ error: 'Internal Server Error' });
   };
+};
+
+module.exports.getBookByid = async(req, res) => {
+  try {
+    const { id } = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) return res.status(400).json({ error: 'Invalid ID format' });
+
+    const book = await BookService.getById(id);
+    return book.error
+      ? res.status(400).json({ error: book.error })
+      : res.status(200).json(book);
+  } catch (error) {
+    res.status(500).json({ error: 'Internal Server Error' });
+  }
 };
 
 module.exports.create = async(req, res) => {
@@ -25,6 +40,8 @@ module.exports.create = async(req, res) => {
 module.exports.update = async(req, res) => {
   try {
     const { id } = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) return res.status(400).json({ error: 'Invalid ID format' });
+
     const book = await BookService.update(id, req.body);
     return book.error
       ? res.status(400).json({ error: book.error })
@@ -37,6 +54,8 @@ module.exports.update = async(req, res) => {
 module.exports.deleteBook = async(req, res) => {
   try {
     const { id } = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) return res.status(400).json({ error: 'Invalid ID format' });
+    
     const book = await BookService.delete(id);
     return book.error
       ? res.status(400).json({ error: book.error })
